requirements.txt
Django==1.9.1
djangorestframework==3.3.2
django-simple-history==1.7.0
wheel==0.24.0
django-oauth-toolkit
markdown
django-filter
django-import-export
django_mobile_app_distribution
Pillow